split classes into own files, add docs
